+++ typescript/ccip-server/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/ccip-server
 
+## 8.9.0
+
 ## 8.8.1
 
 ## 8.8.0